[chore] SwiftLint 규칙 적용 및 파일 분리
