+++ zzimtory/Search/View/ItemCardsView/Shuffle/Classes/Internal/StringUtils.swift
@@ -24,7 +24,7 @@
 
 import UIKit
 
-//swiftlint:disable line_length
+// swiftlint:disable line_length
 enum StringUtils {
 
   static func createInvalidUpdateErrorString(newCount: Int,
@@ -34,4 +34,4 @@ enum StringUtils {
     return "Invalid update: invalid number of cards. The number of cards contained in the card stack after the update (\(newCount)) must be equal to the number of cards contained in the card stack before the update (\(oldCount)), plus or minus the number of cards inserted or deleted (\(insertedCount) inserted, \(deletedCount) deleted)"
   }
 }
-//swiftlint:enable line_length
+// swiftlint:enable line_length
