+++ zzimtory/Search/View/ItemCardsView/Shuffle/Classes/SwipeCardStack/SwipeCardStack.swift
@@ -24,7 +24,8 @@
 
 import UIKit
 
-open class SwipeCardStack: UIView, SwipeCardDelegate, UIGestureRecognizerDelegate {
+open class SwipeCardStack: UIView, SwipeCardDelegate,
+                           UIGestureRecognizerDelegate {
 
   /// A internal structure for a `SwipeCard` and it's corresponding index in the card stack's `dataSource`.
   struct Card {
